moved ProjectListWidget to widgets.py
